feat: melhoria com avrogen + flush por tempo - ainda nao esta consumindo todas as msgs

+++ Domain/Entities/User.cs
@@ -5,6 +5,6 @@
         public string Name { get; set; }
         public int Age { get; set; }
         public int Id { get; set; }
-        public decimal Money { get; set; }
+        public double Money { get; set; }
     }
 }
